+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Fixed
+- Fix joining inside the majority interval
 
 ## [v2.0.3-rc] - 2020-03-19
 ### Fixed